+++ pkg/exchange/bingx/ws_private_test.go
@@ -3,8 +3,8 @@
 package bingx
 
 import (
-	"compress/gzip"
 	"bytes"
+	"compress/gzip"
 	"encoding/json"
 	"sync"
 	"testing"